+++ CHANGES.rst
@@ -41,6 +41,7 @@ Unreleased
         :meth:`Signature.bind` and :func:`inspect.signature` instead.
     -   Remove ``detect_utf_encoding``, it's built-in to ``json.loads``.
     -   Remove ``format_string``, use :class:`string.Template` instead.
+    -   Remove ``escape`` and ``unescape``. Use MarkupSafe instead.
 
 -   Default values passed to ``Headers`` are validated the same way
     values added later are. :issue:`1608`